Math problem of the week script
This script changes changes the images in the math section of the week each week, but for testing purposes the setInterval() function is set to change the image every second. This program uses the images stored in the img2 folder and it goes in the head section of the index.html(home file) file.